+++ desc/objectives/_omnigenity.py
@@ -141,7 +141,7 @@ class QuasisymmetryBoozer(_Objective):
             "resolution for surface averages",
         )
 
-        self._data_keys = ["|B|_mn"]
+        self._data_keys = ["|B|_mn_B"]
 
         timer = Timer()
         if verbose > 0:
@@ -207,7 +207,7 @@ class QuasisymmetryBoozer(_Objective):
             transforms=constants["transforms"],
             profiles=constants["profiles"],
         )
-        B_mn = constants["matrix"] @ data["|B|_mn"]
+        B_mn = constants["matrix"] @ data["|B|_mn_B"]
         return B_mn[constants["idx"]]
 
     @property
@@ -739,7 +739,7 @@ class Omnigenity(_Objective):
             field_grid = self._field_grid
 
         self._dim_f = field_grid.num_nodes
-        self._eq_data_keys = ["|B|_mn"]
+        self._eq_data_keys = ["|B|_mn_B"]
         self._field_data_keys = ["|B|", "theta_B", "zeta_B"]
 
         errorif(
@@ -919,7 +919,7 @@ class Omnigenity(_Objective):
             )
         ).T
         B_eta_alpha = jnp.matmul(
-            constants["eq_transforms"]["B"].basis.evaluate(nodes), eq_data["|B|_mn"]
+            constants["eq_transforms"]["B"].basis.evaluate(nodes), eq_data["|B|_mn_B"]
         )
         omnigenity_error = B_eta_alpha - field_data["|B|"]
         weights = (self.eta_weight + 1) / 2 + (self.eta_weight - 1) / 2 * jnp.cos(